Move common tests to tests module.
