--- packages/library/src/commonTest/kotlin/io/realm/Utils.kt
@@ -1,10 +1,0 @@
-package io.realm
-
-object Utils {
-    fun createRandomString(length: Int): String {
-        val allowedChars = ('A'..'Z') + ('a'..'z') + ('0'..'9')
-        return (1..length)
-            .map { allowedChars.random() }
-            .joinToString("")
-    }
-}
