+++ test/src/androidTest/kotlin/io/realm/ImportTests.kt
@@ -35,7 +35,7 @@ class ImportTests {
 
     @BeforeTest
     fun setup() {
-        tmpDir = Utils.createTempDir()
+        tmpDir = PlatformUtils.createTempDir()
         val configuration =
             RealmConfiguration(path = "$tmpDir/default.realm", schema = setOf(Parent::class, Child::class, Sample::class))
         realm = Realm.open(configuration)
@@ -43,7 +43,7 @@ class ImportTests {
 
     @AfterTest
     fun tearDown() {
-        Utils.deleteTempDir(tmpDir)
+        PlatformUtils.deleteTempDir(tmpDir)
     }
 
     @Test
