+++ test/src/androidTest/kotlin/io/realm/MemoryTests.kt
@@ -40,12 +40,12 @@ class MemoryTests {
     @ExperimentalPathApi
     @Before
     fun setup() {
-        tmpDir = Utils.createTempDir()
+        tmpDir = PlatformUtils.createTempDir()
     }
 
     @After
     fun tearDown() {
-        Utils.deleteTempDir(tmpDir)
+        PlatformUtils.deleteTempDir(tmpDir)
     }
 
     @Test
