+++ test/src/androidTest/kotlin/io/realm/NullabilityTests.kt
@@ -30,14 +30,14 @@ class NullabilityTests {
 
     @BeforeTest
     fun setup() {
-        tmpDir = Utils.createTempDir()
+        tmpDir = PlatformUtils.createTempDir()
         val configuration = RealmConfiguration(path = "$tmpDir/default.realm", schema = setOf(Nullability::class))
         realm = Realm.open(configuration)
     }
 
     @AfterTest
     fun tearDown() {
-        Utils.deleteTempDir(tmpDir)
+        PlatformUtils.deleteTempDir(tmpDir)
     }
 
     @Test
