+++ test/src/commonTest/kotlin/io/realm/SampleTests.kt
@@ -32,7 +32,7 @@ class SampleTests {
 
     @BeforeTest
     fun setup() {
-        tmpDir = Utils.createTempDir()
+        tmpDir = PlatformUtils.createTempDir()
         val configuration = RealmConfiguration(path = "$tmpDir/default.realm", schema = setOf(Sample::class))
         realm = Realm.open(configuration)
     }
@@ -40,7 +40,7 @@ class SampleTests {
     @AfterTest
     fun tearDown() {
         realm.close()
-        Utils.deleteTempDir(tmpDir)
+        PlatformUtils.deleteTempDir(tmpDir)
     }
 
     @Test
