+++ test/src/commonTest/kotlin/io/realm/Utils.kt
@@ -17,7 +17,21 @@
 
 package io.realm
 
-expect object Utils {
+// Platform dependant helper methods
+expect object PlatformUtils {
     fun createTempDir(): String
     fun deleteTempDir(path: String)
 }
+
+// Platform independent helper methods
+object Utils {
+    fun createRandomString(length: Int): String {
+        val allowedChars = ('A'..'Z') + ('a'..'z') + ('0'..'9')
+        return (1..length)
+                .map { allowedChars.random() }
+                .joinToString("")
+    }
+}
+
+
+
